updated the libfuzzer tests

+++ elina_oct/tests/libFuzzer/test_oct.h
@@ -15,14 +15,14 @@
 #define R(i) i < K
 
 bool create_octagon(opt_oct_t* octagon, elina_manager_t* man, opt_oct_t * top,
-		unsigned short int dim, const uint64_t *data, size_t dataSize,
+		unsigned short int dim, const int *data, size_t dataSize,
 		unsigned int *dataIndex);
 
-bool make_fuzzable(void *array, size_t size, const uint64_t *data, size_t dataSize,
+bool make_fuzzable(void *array, size_t size, const int *data, size_t dataSize,
 		unsigned int *dataIndex);
 bool assume_fuzzable(bool condition);
 
-bool make_fuzzable_dimension(size_t * dim, const uint64_t *data, size_t dataSize,
+bool make_fuzzable_dimension(size_t * dim, const int *data, size_t dataSize,
 		unsigned int *dataIndex);
 
 #endif /* TEST_OCT_H_ */
